- applied cleanups and annotations to pass the 'sparse' analysis tests   more information about sparse is at:   http://groups.google.com/groups?hl=en&lr=&selm=2VezS-rN-3%40gated-at.bofh.it
- we still get 'sparse' warnings due to a function that we use for both reading
  data into kernel buffers (via readpage for mmap) and user buffers (for normal
  file_read calls).  without redundant code, there's no trivial way to get
  around the warning, so it remains for now.
- due to preprocessor elimination of the annotations used, this appears to also
  work on 2.4.x kernels and setups that don't understand the annotations
- misc cleanups


SVN revision 4136

+++ src/io/dev/pint-dev-shared.h
@@ -12,11 +12,11 @@
 #define __PINT_DEV_SHARED_H
 
 /* supported ioctls */
-#define PVFS_DEV_GET_MAGIC                      1
-#define PVFS_DEV_GET_MAX_UPSIZE                 2
-#define PVFS_DEV_GET_MAX_DOWNSIZE               3
-#define PVFS_DEV_MAP                            4
-#define PVFS_DEV_REMOUNT_ALL                    5
+#define PVFS_DEV_GET_MAGIC          (unsigned int)1
+#define PVFS_DEV_GET_MAX_UPSIZE     (unsigned int)2
+#define PVFS_DEV_GET_MAX_DOWNSIZE   (unsigned int)3
+#define PVFS_DEV_MAP                (unsigned int)4
+#define PVFS_DEV_REMOUNT_ALL        (unsigned int)5
 
 /* This is the number of discrete buffers we will break the mapped I/O
  * region into.  In some sense it governs the number of concurrent I/O
